+++ ktor-server/ktor-server-core/posix/src/io/ktor/server/engine/EnvironmentUtils.posix.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.server.engine
+
+import kotlinx.cinterop.*
+import platform.posix.*
+
+@OptIn(ExperimentalForeignApi::class)
+internal actual fun getEnvironmentProperty(key: String): String? = getenv(key)?.toKString()
