--- ktor-server/ktor-server-core/web/src/io/ktor/server/request/ApplicationReceiveFunctions.web.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright 2014-2025 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.server.request
-
-@PublishedApi
-internal actual val DEFAULT_FORM_FIELD_LIMIT: Long
-    get() = 50 * 1024 * 1024L
